Internal change
PiperOrigin-RevId: 538022874
