+++ official/projects/yt8m/configs/yt8m.py
@@ -49,6 +49,7 @@ class DataConfig(cfg.DataConfig):
     include_video_id: `True` means include video id (string) in the input to the
       model.
     temporal_stride: Not used. Need to deprecated.
+    sample_random_frames: If sample random frames.
     max_frames: Maxim Number of frames in a input example. It is used to crop
       the input in the temporal dimension.
     num_sample_frames: Number of frames to sample for each input example.
@@ -75,6 +76,7 @@ class DataConfig(cfg.DataConfig):
   include_video_id: bool = False
   temporal_stride: int = 1
   max_frames: int = 300
+  sample_random_frames: bool = True
   num_sample_frames: int = 300  # set smaller to allow random sample (Parser)
   num_classes: int = 3862
   num_devices: int = 1
@@ -100,26 +102,63 @@ def yt8m(is_training):
 
 
 @dataclasses.dataclass
-class MoeModel(hyperparams.Config):
-  """The model config."""
-  num_mixtures: int = 5
-  use_input_context_gate: bool = False
-  use_output_context_gate: bool = False
-  vocab_as_last_dim: bool = False
-
-
-@dataclasses.dataclass
 class DbofModel(hyperparams.Config):
   """The model config."""
   cluster_size: int = 3000
   hidden_size: int = 2000
   add_batch_norm: bool = True
-  sample_random_frames: bool = True
+  pooling_method: str = 'average'
   use_context_gate_cluster_layer: bool = False
   context_gate_cluster_bottleneck_size: int = 0
-  pooling_method: str = 'average'
-  agg_classifier_model: str = 'MoeModel'
-  agg_model: hyperparams.Config = MoeModel()
+
+
+@dataclasses.dataclass
+class Backbone(hyperparams.OneOfConfig):
+  """Configuration for backbones.
+
+  Attributes:
+    type: 'str', type of backbone be used, one of the fields below.
+    dbof: dbof backbone config.
+  """
+  type: Optional[str] = None
+  dbof: DbofModel = DbofModel()
+
+
+@dataclasses.dataclass
+class MoeModel(hyperparams.Config):
+  """The MoE model config."""
+
+  num_mixtures: int = 5
+  vocab_as_last_dim: bool = False
+  use_input_context_gate: bool = False
+  use_output_context_gate: bool = False
+
+
+@dataclasses.dataclass
+class LogisticModel(hyperparams.Config):
+  """The logistic model config."""
+  return_logits: bool = False
+
+
+@dataclasses.dataclass
+class Head(hyperparams.OneOfConfig):
+  """Configuration for aggreagation heads.
+
+  Attributes:
+    type: 'str', type of head be used, one of the fields below.
+    moe: MoE head config.
+    logistic: Logistic head config.
+  """
+  type: Optional[str] = None
+  moe: MoeModel = MoeModel()
+  logistic: LogisticModel = LogisticModel()
+
+
+@dataclasses.dataclass
+class VideoClassificationModel(hyperparams.Config):
+  """The classifier model config."""
+  backbone: Backbone = Backbone(type='dbof')
+  head: Head = Head(type='moe')
   norm_activation: common.NormActivation = common.NormActivation(
       activation='relu', use_sync_bn=False)
 
@@ -146,7 +185,7 @@ class Evaluation(hyperparams.Config):
 @dataclasses.dataclass
 class YT8MTask(cfg.TaskConfig):
   """The task config."""
-  model: DbofModel = DbofModel()
+  model: VideoClassificationModel = VideoClassificationModel()
   train_data: DataConfig = yt8m(is_training=True)
   validation_data: DataConfig = yt8m(is_training=False)
   losses: Losses = Losses()
