+++ official/projects/yt8m/modeling/yt8m_model.py
@@ -15,23 +15,25 @@
 """YT8M prediction model definition."""
 
 import functools
-from typing import Any, Optional
+from typing import Any
 
 from absl import logging
 import tensorflow as tf
 
 from official.projects.yt8m.configs import yt8m as yt8m_cfg
-from official.projects.yt8m.modeling import nn_layers
+from official.projects.yt8m.modeling import backbones  # pylint: disable=unused-import
+from official.projects.yt8m.modeling import heads
+from official.vision.modeling.backbones import factory
 
 
 layers = tf.keras.layers
 
 
-class DbofModel(tf.keras.Model):
-  """A YT8M model class builder.
+class VideoClassificationModel(tf.keras.Model):
+  """A video classification model class builder.
 
-  Creates a Deep Bag of Frames model.
-  The model projects the features for each frame into a higher dimensional
+  The model consists of a backbone (dbof) and a classification head.
+  The dbof backbone projects features for each frame into a higher dimensional
   'clustering' space, pools across frames in that space, and then
   uses a configurable video-level model to classify the now aggregated features.
   The model will randomly sample either frames or sequences of frames during
@@ -40,17 +42,20 @@ class DbofModel(tf.keras.Model):
 
   def __init__(
       self,
-      params: yt8m_cfg.DbofModel,
+      params: yt8m_cfg.VideoClassificationModel,
+      backbone: tf.keras.Model | None = None,
       num_classes: int = 3862,
       input_specs: layers.InputSpec = layers.InputSpec(
-          shape=[None, None, 1152]),
-      l2_weight_decay: Optional[float] = None,
+          shape=[None, None, 1152]
+      ),
+      l2_weight_decay: float | None = None,
       **kwargs,
   ):
-    """YT8M Dbof model initialization function.
+    """YT8M video classification model initialization function.
 
     Args:
-      params: model configuration parameters
+      params: Model configuration parameters.
+      backbone: Optional backbone model. Will build a backbone if None.
       num_classes: `int` number of classes in dataset.
       input_specs: `tf.keras.layers.InputSpec` specs of the input tensor.
         [batch_size x num_frames x num_features]
@@ -69,33 +74,57 @@ class DbofModel(tf.keras.Model):
         "l2_weight_decay": l2_weight_decay,
     }
 
-    self.dbof_backbone = nn_layers.Dbof(
-        params,
-        num_classes,
-        input_specs,
-        l2_weight_decay,
-        **kwargs,
+    if backbone is None:
+      # Divide weight decay by 2.0 to match the implementation of tf.nn.l2_loss.
+      # (https://www.tensorflow.org/api_docs/python/tf/keras/regularizers/l2)
+      # (https://www.tensorflow.org/api_docs/python/tf/nn/l2_loss)
+      l2_regularizer = (
+          tf.keras.regularizers.l2(l2_weight_decay / 2.0)
+          if l2_weight_decay
+          else None
+      )
+      backbone = factory.build_backbone(
+          input_specs=input_specs,
+          backbone_config=params.backbone,
+          norm_activation_config=params.norm_activation,
+          l2_regularizer=l2_regularizer,
+          **kwargs,
+      )
+
+    self.backbone = backbone
+    self.build_head()
+
+  def build_head(self):
+    logging.info("Build DbofModel with %s.", self._params.head.type)
+    head_cfg = self._params.head.get()
+    if self._params.head.type == "moe":
+      normalizer_params = dict(
+          synchronized=self._params.norm_activation.use_sync_bn,
+          momentum=self._params.norm_activation.norm_momentum,
+          epsilon=self._params.norm_activation.norm_epsilon,
+      )
+      aggregation_head = functools.partial(
+          heads.MoeModel, normalizer_params=normalizer_params
+      )
+    elif self._params.head.type == "logistic":
+      aggregation_head = heads.LogisticModel
+    else:
+      logging.warn("Skip build head type: %s", self._params.head.type)
+      return
+
+    l2_regularizer = (
+        tf.keras.regularizers.l2(self._l2_weight_decay / 2.0)
+        if self._l2_weight_decay
+        else None
     )
-
-    logging.info("Build DbofModel with %s.", params.agg_classifier_model)
-    if hasattr(nn_layers, params.agg_classifier_model):
-      aggregation_head = getattr(nn_layers, params.agg_classifier_model)
-      if params.agg_classifier_model == "MoeModel":
-        normalizer_params = dict(
-            synchronized=params.norm_activation.use_sync_bn,
-            momentum=params.norm_activation.norm_momentum,
-            epsilon=params.norm_activation.norm_epsilon,
-        )
-        aggregation_head = functools.partial(
-            aggregation_head, normalizer_params=normalizer_params)
-
-      if params.agg_model is not None:
-        kwargs.update(params.agg_model.as_dict())
-      self.head = aggregation_head(
-          input_specs=layers.InputSpec(shape=[None, params.hidden_size]),
-          vocab_size=num_classes,
-          l2_penalty=l2_weight_decay,
-          **kwargs)
+    self.head = aggregation_head(
+        input_specs=layers.InputSpec(
+            shape=[None, self._params.backbone.get().hidden_size]
+        ),
+        vocab_size=self._num_classes,
+        l2_regularizer=l2_regularizer,
+        **head_cfg.as_dict(),
+    )
 
   def get_config(self):
     return self._config_dict
@@ -106,7 +135,12 @@ class DbofModel(tf.keras.Model):
 
   def call(
       self, inputs: tf.Tensor, training: Any = None, mask: Any = None
-  ) -> tf.Tensor:
-    features = self.dbof_backbone(inputs)
+  ) -> dict[str, tf.Tensor]:
+    features = self.backbone(inputs)
     outputs = self.head(features)
-    return outputs["predictions"]
+    return outputs
+
+  @property
+  def checkpoint_items(self) -> dict[str, Any]:
+    """Returns a dictionary of items to be additionally checkpointed."""
+    return dict(backbone=self.backbone, head=self.head)
