+++ official/projects/yt8m/modeling/yt8m_model_test.py
@@ -37,21 +37,21 @@ class YT8MNetworkTest(parameterized.TestCase, tf.test.TestCase):
     input_specs = tf.keras.layers.InputSpec(shape=[None, None, feature_dims])
 
     num_classes = 3862
-    model = yt8m_model.DbofModel(
+    model = yt8m_model.VideoClassificationModel(
         params=yt8m_cfg.YT8MTask.model,
         num_classes=num_classes,
         input_specs=input_specs)
 
     # batch = 2 -> arbitrary value for test.
     inputs = np.random.rand(2, num_frames, feature_dims)
-    logits = model(inputs)
-    self.assertAllEqual([2, num_classes], logits.numpy().shape)
+    predictions = model(inputs)['predictions']
+    self.assertAllEqual([2, num_classes], predictions.numpy().shape)
 
   def test_serialize_deserialize(self):
-    model = yt8m_model.DbofModel(params=yt8m_cfg.YT8MTask.model)
+    model = yt8m_model.VideoClassificationModel(params=yt8m_cfg.YT8MTask.model)
 
     config = model.get_config()
-    new_model = yt8m_model.DbofModel.from_config(config)
+    new_model = yt8m_model.VideoClassificationModel.from_config(config)
 
     # If the serialization was successful,
     # the new config should match the old.
