+++ official/projects/yt8m/tasks/yt8m_task.py
@@ -25,8 +25,8 @@ from official.modeling import tf_utils
 from official.projects.yt8m.configs import yt8m as yt8m_cfg
 from official.projects.yt8m.dataloaders import yt8m_input
 from official.projects.yt8m.eval_utils import eval_util
+from official.projects.yt8m.modeling import yt8m_model
 from official.projects.yt8m.modeling import yt8m_model_utils as utils
-from official.projects.yt8m.modeling.yt8m_model import DbofModel
 
 
 @task_factory.register_task_cls(yt8m_cfg.YT8MTask)
@@ -45,7 +45,7 @@ class YT8MTask(base_task.Task):
     l2_weight_decay = self.task_config.losses.l2_weight_decay
     # Model configuration.
     model_config = self.task_config.model
-    model = DbofModel(
+    model = yt8m_model.VideoClassificationModel(
         params=model_config,
         input_specs=input_specs,
         num_classes=train_cfg.num_classes,
@@ -53,12 +53,11 @@ class YT8MTask(base_task.Task):
 
     non_trainable_batch_norm_variables = []
     non_trainable_extra_variables = []
-    for var in model.variables:
-      if not var.trainable:
-        if 'moving_mean' or 'moving_variance' in var.name:
-          non_trainable_batch_norm_variables.append(var)
-        else:
-          non_trainable_extra_variables.append(var)
+    for var in model.non_trainable_variables:
+      if 'moving_mean' in var.name or 'moving_variance' in var.name:
+        non_trainable_batch_norm_variables.append(var)
+      else:
+        non_trainable_extra_variables.append(var)
 
     logging.info(
         'Trainable model variables:\n%s',
@@ -131,7 +130,7 @@ class YT8MTask(base_task.Task):
 
     Args:
       labels: tensor containing truth labels.
-      model_outputs: output logits of the classifier.
+      model_outputs: output probabilities of the classifier.
       label_weights: optional tensor of label weights.
       aux_losses: tensor containing auxiliarly loss tensors, i.e. `losses` in
         keras.Model.
@@ -251,7 +250,7 @@ class YT8MTask(base_task.Task):
     # sample random frames / random sequence.
     num_frames = tf.cast(num_frames, tf.float32)
     num_sample_frames = data_config.num_sample_frames
-    if self.task_config.model.sample_random_frames:
+    if data_config.sample_random_frames:
       features = utils.sample_random_frames(
           features, num_frames, num_sample_frames)
     else:
@@ -270,7 +269,6 @@ class YT8MTask(base_task.Task):
     return labels, label_weights
 
   def _postprocess_outputs(self,
-                           inputs,
                            outputs,
                            labels,
                            label_weights,
@@ -282,7 +280,7 @@ class YT8MTask(base_task.Task):
       # remove padding
       outputs = outputs[~tf.reduce_all(labels == -1, axis=1)]
       labels = labels[~tf.reduce_all(labels == -1, axis=1)]
-    return inputs, outputs, labels, label_weights
+    return outputs, labels, label_weights
 
   def train_step(self, inputs, model, optimizer, metrics=None):
     """Does forward and backward.
@@ -303,13 +301,13 @@ class YT8MTask(base_task.Task):
 
     num_replicas = tf.distribute.get_strategy().num_replicas_in_sync
     with tf.GradientTape() as tape:
-      outputs = model(model_inputs, training=True)
+      outputs = model(model_inputs, training=True)['predictions']
       # Casting output layer as float32 is necessary when mixed_precision is
       # mixed_float16 or mixed_bfloat16 to ensure output is casted as float32.
       outputs = tf.nest.map_structure(lambda x: tf.cast(x, tf.float32), outputs)
       # Post-process model / label outputs.
-      inputs, outputs, labels, label_weights = self._postprocess_outputs(
-          inputs, outputs, labels, label_weights, training=True)
+      outputs, labels, label_weights = self._postprocess_outputs(
+          outputs, labels, label_weights, training=True)
 
       # Computes per-replica loss
       all_losses = self.build_losses(
@@ -368,10 +366,10 @@ class YT8MTask(base_task.Task):
     model_inputs = self._preprocess_model_inputs(inputs, training=False)
     labels, label_weights = self._preprocess_labels(inputs, training=False)
 
-    outputs = self.inference_step(model_inputs, model)
+    outputs = self.inference_step(model_inputs, model)['predictions']
     outputs = tf.nest.map_structure(lambda x: tf.cast(x, tf.float32), outputs)
-    inputs, outputs, labels, label_weights = self._postprocess_outputs(
-        inputs, outputs, labels, label_weights, training=False)
+    outputs, labels, label_weights = self._postprocess_outputs(
+        outputs, labels, label_weights, training=False)
 
     all_losses = self.build_losses(
         labels=labels,
